Refactored clinical trial fetching into multiple files for readability and organization
